Fix unhandled exception if http request returns null
Angular's Http class will sometimes return null, rather than raising
an error in the observable. This usually happens if CORS preflight checks
fail and the request is rejected by the underlying browser.

This change makes ManUp resolve in the case that http returns null.

Under the hood also making more use of async/await to make the code a bit
nicer to read.